Refactored common code in check_lib.py
